+++ install-dev.sh
@@ -1,14 +1,14 @@
 #!/bin/bash
 
-#Creates symlinks to your working project folder in the nemo-python plugins directory
+INSTALL_ROOT="$HOME/.local/share/nemo-rsync"
+
+mkdir -p "$INSTALL_ROOT"
+
+ln -sfn "$PWD/backend" "$INSTALL_ROOT/backend"
+ln -sfn "$PWD/extension" "$INSTALL_ROOT/extension"
+
 mkdir -p ~/.local/share/nemo-python/extensions
 
-ln -sf \
-"$PWD/extension/nemo_rsync_extension.py" \
-~/.local/share/nemo-python/extensions/
-
-mkdir -p ~/.local/share/nemo-rsync
-
 ln -sfn \
-"$PWD/backend" \
-~/.local/share/nemo-rsync/backend
+"$INSTALL_ROOT/extension/nemo_rsync_extension.py" \
+~/.local/share/nemo-python/extensions/nemo_rsync_extension.py